+++ dart/simulation/experimental/frame/fixed_frame.cpp
@@ -64,9 +64,7 @@ void FixedFrame::setLocalTransform(const Eigen::Isometry3d& transform)
   auto& props = registry.get<comps::FixedFrameProperties>(m_entity);
   props.localTransform = transform;
 
-  // Invalidate cache
-  auto& cache = registry.get<comps::FrameCache>(m_entity);
-  cache.needTransformUpdate = true;
+  markSubtreeTransformCacheDirty();
 }
 
 //==============================================================================